Revert "call static rlp class"
This reverts commit c98746039409b36119e1dd7b91f9084dcad33826.
